requirements.txt
asgiref==3.2.2
atomicwrites==1.3.0
attrs==19.1.0
branca==0.3.1
certifi==2019.6.16
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
dj-stripe==2.1.1
Django==2.2.7
django-filter==2.2.0
django-registration==3.0.1
django-settings==1.3.12
djangorestframework==3.10.3
Flask==1.1.1
Flask-PyMongo==2.3.0
folium==0.10.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.2
packaging==19.1
plotly==4.2.0
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
stripe==2.41.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
whitenoise==4.1.4
zipp==0.6.0
